pyproject.toml
[project]
name = "qdrant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.2",
]
